refactor and remove no necessary code
